+++ js/main.js
@@ -1,4 +1,6 @@
-import './form.js'
+import './form.js';
+import './validation.js'
+import {filtreRooms} from './validation.js';
 import {getDisablePage} from './util.js';
 import {createFetch} from './create-fetch.js';
 import {mapModule} from './map.js';
@@ -6,6 +8,7 @@ import {setFilterAction} from './filters.js';
 import {showAlert} from './alert.js';
 
 getDisablePage();
+filtreRooms(document.querySelectorAll('#capacity option'), 1);
 
 createFetch()
   .then((offers) => {
